Apply textfield properties and object type operator
The Textfield was pushing the cursor to the end of the value in the text field. This problem is exhibited when embedded in a table row, after experimentation, adding a height property to the textfield resolved this, as opposed to not placing the textfield inside the row. A few minor fixes when performing operations and using type based lookups against osname.